+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Added
 - Added retry support (#1616)
 - Added support for VRANGE command (#1623)
+- Added support for idempotent stream API (#1632)
 - Added support for HOTKEYS container command (#1630)
 
 ### Maintenance